Refactor handlers, fix bugs, and remove inconsistencies

+++ grader_service/handlers/roles.py
@@ -1,6 +1,6 @@
 from http import HTTPStatus
 
-import tornado
+from sqlalchemy.orm.exc import ObjectDeletedError
 from tornado.escape import json_decode
 from tornado.web import HTTPError
 
@@ -10,7 +10,8 @@ from grader_service.orm import User
 from grader_service.orm.takepart import Role, Scope
 from grader_service.registry import VersionSpecifier, register_handler
 
-@register_handler(r"\/api\/users\/(?P<username>[^\/]+)\/roles", VersionSpecifier.ALL)
+
+@register_handler(r"\/api\/users\/(?P<username>[^\/]+)\/roles\/?", VersionSpecifier.ALL)
 class RoleUserHandler(GraderBaseHandler):
     """
     Tornado Handler class for http requests to /user/{username}/roles.
@@ -35,7 +36,8 @@ class RoleUserHandler(GraderBaseHandler):
         self.set_status(HTTPStatus.OK)
         self.write_json(roles)
 
-@register_handler(r"\/api\/lectures\/(?P<lecture_id>\d*)\/roles", VersionSpecifier.ALL)
+
+@register_handler(r"\/api\/lectures\/(?P<lecture_id>\d*)\/roles\/?", VersionSpecifier.ALL)
 class RoleBaseHandler(GraderBaseHandler):
     """
     Tornado Handler class for http requests to /lectures/{lecture_id}/roles.
@@ -51,10 +53,11 @@ class RoleBaseHandler(GraderBaseHandler):
         """
         lecture_id = parse_ids(lecture_id)
         self.validate_parameters()
+
         roles = self.session.query(Role).filter(Role.lectid == lecture_id).all()
 
         self.set_status(HTTPStatus.OK)
-        self.write_json(roles)
+        self.write_json([r.serialize_with_user() for r in roles])
 
     @authorize([Scope.admin])
     async def post(self, lecture_id: int):
@@ -71,11 +74,15 @@ class RoleBaseHandler(GraderBaseHandler):
 
         :param lecture_id: id of the lecture
         :type lecture_id: int
-        raises HTTPError: throws err if one user was not found
+        :raises HTTPError: throws err if one user was not found
         """
         lecture_id = parse_ids(lecture_id)
         self.validate_parameters()
-        body = tornado.escape.json_decode(self.request.body)
+        body = json_decode(self.request.body)
+
+        lecture = self.get_lecture(lecture_id)
+        if lecture is None:
+            raise HTTPError(HTTPStatus.NOT_FOUND, reason="Lecture not found")
 
         roles = []
         for user_req in body["users"]:
@@ -99,39 +106,51 @@ class RoleBaseHandler(GraderBaseHandler):
         self.session.commit()
 
         self.set_status(HTTPStatus.CREATED)
-        self.write_json(roles)
+        self.write_json([r.serialize_with_user() for r in roles])
 
     @authorize([Scope.admin])
     async def delete(self, lecture_id: int):
         """
         Deletes roles for a specific lecture.
 
-        Request body example:
-        {
-            "users": ["alice", "bob"]
-        }
+        Query parameter example:
+        ?usernames=alice,boby
 
         :param lecture_id: id of the lecture
         :type lecture_id: int
-        raises HTTPError: throws err if one user was not found
+        :raises HTTPError: if the lecture does not exist, no usernames were provided,
+                           or at least one user cannot be found
         """
         lecture_id = parse_ids(lecture_id)
-        self.validate_parameters()
-        body = tornado.escape.json_decode(self.request.body)
+        self.validate_parameters("usernames")
+        raw_usernames = self.get_argument("usernames", "")
 
-        roles = []
-        for username in body['users']:
+        try:
+            # Roles can not be soft-deleted
+            if not self.user.is_admin:
+                raise HTTPError(HTTPStatus.FORBIDDEN, reason="Only Admins can delete roles.")
 
-            user = self.session.query(User).filter(User.name == username).one_or_none()
-            if user is None:
-                self.session.rollback()
-                raise HTTPError(HTTPStatus.NOT_FOUND, reason=f"User {username} not found")
+            lecture = self.get_lecture(lecture_id)
+            if lecture is None:
+                raise HTTPError(HTTPStatus.NOT_FOUND, reason="Lecture not found")
 
-            role = self.session.query(Role) \
-                .filter(Role.user_id == user.id) \
-                .filter(Role.lectid == lecture_id) \
-                .one_or_none()
-            if role is not None:
-                roles.append(role)
-                self.session.delete(role)
-        self.session.commit()
+            if len(raw_usernames.strip()) == 0:
+                raise HTTPError(HTTPStatus.BAD_REQUEST, reason="usernames cannot be empty")
+            usernames = raw_usernames.split(",")
+
+            for username in usernames:
+                user = self.session.query(User).filter(User.name == username).one_or_none()
+                if user is None:
+                    self.session.rollback()
+                    raise HTTPError(HTTPStatus.NOT_FOUND, reason=f"User {username} was not found")
+
+                role = self.session.query(Role) \
+                    .filter(Role.user_id == user.id) \
+                    .filter(Role.lectid == lecture_id) \
+                    .one_or_none()
+                if role is not None:
+                    self.session.delete(role)
+            self.session.commit()
+        except ObjectDeletedError:
+            raise HTTPError(HTTPStatus.NOT_FOUND, reason="Roles was not found")
+        self.write("OK")
